updated callbackUrl to env url and moved authOptions to lib auth.ts file to fix deployment issues
